refactor: share a single HttpClient across application, reorganise some networking
