+++ src/XIVLauncher.Common.Unix/Compatibility/CompatibilityTools.cs
@@ -78,25 +78,24 @@ public class CompatibilityTools
             wineSettings.Prefix.Create();
     }
 
-    public async Task EnsureTool(DirectoryInfo tempPath)
+    public async Task EnsureTool(HttpClient httpClient, DirectoryInfo tempPath)
     {
         if (!File.Exists(Wine64Path))
         {
             Log.Information($"Compatibility tool does not exist, downloading {Settings.WineRelease.DownloadUrl}");
-            await DownloadTool(tempPath).ConfigureAwait(false);
+            await DownloadTool(httpClient, tempPath).ConfigureAwait(false);
         }
 
         EnsurePrefix();
-        await Dxvk.Dxvk.InstallDxvk(Settings.Prefix, dxvkDirectory, dxvkVersion).ConfigureAwait(false);
+        await Dxvk.Dxvk.InstallDxvk(httpClient, Settings.Prefix, dxvkDirectory, dxvkVersion).ConfigureAwait(false);
 
         IsToolReady = true;
     }
 
-    private async Task DownloadTool(DirectoryInfo tempPath)
+    private async Task DownloadTool(HttpClient httpClient, DirectoryInfo tempPath)
     {
-        using var client = HappyEyeballsHttp.CreateHttpClient();
         var tempFilePath = Path.Combine(tempPath.FullName, $"{Guid.NewGuid()}");
-        await File.WriteAllBytesAsync(tempFilePath, await client.GetByteArrayAsync(Settings.WineRelease.DownloadUrl).ConfigureAwait(false)).ConfigureAwait(false);
+        await File.WriteAllBytesAsync(tempFilePath, await httpClient.GetByteArrayAsync(Settings.WineRelease.DownloadUrl).ConfigureAwait(false)).ConfigureAwait(false);
         if (!CompatUtil.EnsureChecksumMatch(tempFilePath, Settings.WineRelease.Checksums))
         {
             throw new InvalidDataException("SHA512 checksum verification failed");
