+++ src/XIVLauncher.Common.Unix/Compatibility/Dxvk/Dxvk.cs
@@ -41,7 +41,7 @@ public enum DxvkHudType
 
 public static class Dxvk
 {
-    public static async Task InstallDxvk(DirectoryInfo prefix, DirectoryInfo installDirectory, DxvkVersion version)
+    public static async Task InstallDxvk(HttpClient httpClient, DirectoryInfo prefix, DirectoryInfo installDirectory, DxvkVersion version)
     {
         if (version is DxvkVersion.Disabled)
         {
@@ -59,7 +59,7 @@ public static class Dxvk
         if (!Directory.Exists(dxvkPath))
         {
             Log.Information("DXVK does not exist, downloading");
-            await DownloadDxvk(installDirectory, release.DownloadUrl, release.Checksum).ConfigureAwait(false);
+            await DownloadDxvk(httpClient, installDirectory, release.DownloadUrl, release.Checksum).ConfigureAwait(false);
         }
 
         var system32 = Path.Combine(prefix.FullName, "drive_c", "windows", "system32");
@@ -71,12 +71,11 @@ public static class Dxvk
         }
     }
 
-    private static async Task DownloadDxvk(DirectoryInfo installDirectory, string url, string checksum)
+    private static async Task DownloadDxvk(HttpClient httpClient, DirectoryInfo installDirectory, string url, string checksum)
     {
-        using var client = HappyEyeballsHttp.CreateHttpClient();
         var tempPath = PlatformHelpers.GetTempFileName();
 
-        File.WriteAllBytes(tempPath, await client.GetByteArrayAsync(url).ConfigureAwait(false));
+        File.WriteAllBytes(tempPath, await httpClient.GetByteArrayAsync(url).ConfigureAwait(false));
 
         if (!CompatUtil.EnsureChecksumMatch(tempPath, [checksum]))
         {
