+++ src/XIVLauncher.Core/Net/LauncherClientConfig.cs
@@ -1,9 +1,9 @@
 
-using Serilog;
-
 using System.Net.Http.Json;
 
-namespace XIVLauncher.Core;
+using Serilog;
+
+namespace XIVLauncher.Core.Net;
 
 
 /// <summary>
