+++ src/XIVLauncher.Core/Program.cs
@@ -1,11 +1,11 @@
+using System.Numerics;
+using System.Runtime.InteropServices;
+
 using Config.Net;
 
 using Hexa.NET.SDL3;
 
 using Serilog;
-
-using System.Numerics;
-using System.Runtime.InteropServices;
 
 using XIVLauncher.Common;
 using XIVLauncher.Common.Dalamud;
@@ -23,6 +23,7 @@ using XIVLauncher.Core.Accounts.Secrets.Providers;
 using XIVLauncher.Core.Components.LoadingPage;
 using XIVLauncher.Core.Configuration;
 using XIVLauncher.Core.Configuration.Parsers;
+using XIVLauncher.Core.Net;
 using XIVLauncher.Core.Style;
 
 using SDLGPUDevice = Hexa.NET.SDL3.SDLGPUDevice;
@@ -49,10 +50,7 @@ sealed class Program
     public static DalamudOverlayInfoProxy DalamudLoadInfo { get; private set; } = null!;
     public static CompatibilityTools CompatibilityTools { get; private set; } = null!;
     public static ISecretProvider Secrets { get; private set; } = null!;
-    public static HttpClient HttpClient { get; private set; } = new()
-    {
-        Timeout = TimeSpan.FromSeconds(5)
-    };
+    public static HttpClient HttpClient { get; private set; } = HappyEyeballsHttp.CreateHttpClient();
     public static PatchManager? Patcher { get; set; } = null;
     public static Storage storage = null!;
     public static DirectoryInfo DotnetRuntime => storage.GetFolder("runtime");
